Update tsv-to-csv.py
Corrected a ParserError from pandas that causes the script to crash when an escape character like a loose double-quote is encountered. The issue doesn't happen for double-quote pairs in a column, only a loose one.
In this case, I simply added a routine to backup the file to an "errors" directory and skip the remainder of the file.